\name{vim.buildomnils}
\alias{vim.buildomnils}
\title{
Build omnilist for Vim.
}
\description{
Build the omnils file of one library for Vim. The function is not supposed to be
called by the user. 
}
\usage{
vim.buildomnils(p)
}
\arguments{
  \item{p}{Package name.}
}
\author{ Jakson Alves de Aquino <jalvesaq@gmail.com> }
