
\name{vim.openpdf}
\alias{vim.openpdf}
\title{Run the pdf file in external application}
\description{
  Open the pdf file in external application.
}
\usage{
vim.openpdf(x, quiet = FALSE)
}
\arguments{
  \item{x}{The file name.}
  \item{quiet}{Whether to suppress output of application opening the PDF.}
}
